+++ src/AhRulesBot/AhRulesBot/AhRulesBot.csproj
@@ -7,6 +7,7 @@
 
   <ItemGroup>
     <PackageReference Include="CsvHelper" Version="26.0.1" />
+    <PackageReference Include="Google.Apis.Docs.v1" Version="1.51.0.2252" />
     <PackageReference Include="Google.Apis.Drive.v3" Version="1.51.0.2265" />
     <PackageReference Include="Microsoft.Extensions.Hosting" Version="5.0.0" />
     <PackageReference Include="Serilog" Version="2.10.0" />
