+++ src/AhRulesBot/AhRulesBot/BotRequestsProcessing/Handlers/CommandMessageHandler.cs
@@ -4,9 +4,9 @@ using System.Collections.Generic;
 
 namespace AhRulesBot.BotRequestsProcessing.Handlers
 {
-    public class CommandMessageHandler : IMessageHandler
+    internal class CommandMessageHandler : IMessageHandler
     {
-        private IMessageHandler _next;
+        private readonly IMessageHandler _next;
 
         public CommandMessageHandler(IMessageHandler next)
         {
