+++ src/AhRulesBot/AhRulesBot/BotRequestsProcessing/Handlers/LongMessageHandler.cs
@@ -4,7 +4,7 @@ using System.Linq;
 
 namespace AhRulesBot.BotRequestsProcessing.Handlers
 {
-    public class LongMessageHandler : IMessageHandler
+    internal class LongMessageHandler : IMessageHandler
     {
         private IMessageHandler _next;
