+++ src/AhRulesBot/AhRulesBot/BotRequestsProcessing/Interfaces/ITelegramMessageProcessor.cs
@@ -4,7 +4,7 @@ using Telegram.Bot.Types;
 
 namespace AhRulesBot.BotRequestsProcessing.Interfaces
 {
-    public interface ITelegramMessageProcessor
+    internal interface ITelegramMessageProcessor
     {
         public Task Process(Message message, CancellationToken cancellationToken);
     }
