+++ src/AhRulesBot/AhRulesBot/CustomRulesImport/CustomRulesWorker.cs
@@ -5,34 +5,32 @@ using System.Threading;
 using System.Threading.Tasks;
 using AhRulesBot.Models;
 using System.IO;
-using Google.Apis.Auth.OAuth2;
-using Google.Apis.Docs.v1;
 using System.Collections.Generic;
-using Google.Apis.Services;
-using Google.Apis.Drive.v3;
 using AhRulesBot.Infrastructure;
 using System.Linq;
 using System.Globalization;
-using Newtonsoft.Json;
-using Newtonsoft.Json.Serialization;
 using CsvHelper;
 using System.Threading.Channels;
+using AhRulesBot.CustomRulesImport;
 
 namespace AhRulesBot.OldMessagesProcessing
 {
     internal class CustomRulesWorker : BackgroundService
     {
         private readonly ILogger _logger;
+        private readonly RulesDriveService _rulesDriveService;
         private readonly ChannelWriter<List<CustomRuleItem>> _channel;
         private readonly AppConfig _config;
 
         public CustomRulesWorker(
             ILogger logger,
             AppConfig config,
+            RulesDriveService rulesDriveService,
             ChannelWriter<List<CustomRuleItem>> channel)
         {
             _channel = channel;
             _logger = logger;
+            _rulesDriveService = rulesDriveService;
             _config = config;
         }
 
@@ -42,8 +40,7 @@ namespace AhRulesBot.OldMessagesProcessing
             {
                 try
                 {
-                    var drive = GetDrive(GetCredentialParameters());
-                    var customRules = await GetCutomRulesFromDrive(drive, cancellationToken);
+                    var customRules = await GetCutomRulesFromDrive(cancellationToken);
                     await _channel.WriteAsync(customRules, cancellationToken);
 
                     await Task.Delay(TimeSpan.FromHours(5));
@@ -61,41 +58,15 @@ namespace AhRulesBot.OldMessagesProcessing
             _logger.Information("stopping {name}", nameof(CustomRulesWorker));
         }
 
-        private async Task<List<CustomRuleItem>> GetCutomRulesFromDrive(DriveService drive, CancellationToken cancellationToken)
+        private async Task<List<CustomRuleItem>> GetCutomRulesFromDrive(CancellationToken cancellationToken)
         {
-            var googleFileStream = await drive.Files.Export(_config.GoogleFileId, "text/csv").ExecuteAsStreamAsync(cancellationToken);
+            var googleFileStream = await _rulesDriveService.Files.Export(_config.GoogleFileId, "text/csv")
+                .ExecuteAsStreamAsync(cancellationToken);
+
             using var reader = new StreamReader(googleFileStream);
             using var csv = new CsvReader(reader, CultureInfo.InvariantCulture);
+
             return csv.GetRecords<CustomRuleItem>().ToList();
-        }
-
-        private DriveService GetDrive(JsonCredentialParameters credParams)
-        {
-            var accountCred = new ServiceAccountCredential(
-               new ServiceAccountCredential.Initializer(credParams.ClientEmail)
-               {
-                   Scopes = new string[] { DocsService.Scope.DriveReadonly }
-               }
-               .FromPrivateKey(credParams.PrivateKey));
-
-            return new DriveService(new BaseClientService.Initializer()
-            {
-                HttpClientInitializer = accountCred,
-                ApplicationName = "custom-ah-rules",
-            });
-        }
-
-        private JsonCredentialParameters GetCredentialParameters()
-        {
-            var settings = new JsonSerializerSettings
-            {
-                ContractResolver = new DefaultContractResolver
-                {
-                    NamingStrategy = new SnakeCaseNamingStrategy { ProcessDictionaryKeys = true }
-                },
-                Formatting = Formatting.Indented
-            };
-            return JsonConvert.DeserializeObject<JsonCredentialParameters>(File.ReadAllText(_config.GoogleCredFilePath), settings);
         }
     }
 }
