+++ src/AhRulesBot/AhRulesBot/CustomRulesImport/Interfaces/IDriveService.cs
@@ -1,0 +1,9 @@
+﻿using Google.Apis.Drive.v3;
+
+namespace AhRulesBot.BotRequestsProcessing.Interfaces
+{
+    internal interface IDriveService
+    {
+        public FilesResource Files { get; }
+    }
+}
