+++ src/AhRulesBot/AhRulesBot/Infrastructure/AppConfig.cs
@@ -5,10 +5,14 @@ namespace AhRulesBot.Infrastructure
     {
         public long TestChatId { get; set; }
         public long AHChatId { get; set; }
-        public int BotAdminId { get; set; }
+        public long BotAdminId { get; set; }
+
         public string ApiKey { get; set; }
         public string BotName { get; set; }
+
         public string GoogleCredFilePath { get; set; }
         public string GoogleFileId { get; set; }
+
+        public string RulesFilePath { get; set; }
     }
 }
