--- src/AhRulesBot/AhRulesBot/Infrastructure/ServiceCollectionExtensions.cs
@@ -1,94 +1,0 @@
-﻿using AhRulesBot.BotRequestsProcessing.Handlers;
-using AhRulesBot.MessageProcessing.Interfaces;
-using AhRulesBot.Models;
-using Microsoft.Extensions.DependencyInjection;
-using Serilog;
-using Serilog.Events;
-using Serilog.Formatting.Compact;
-using System.Collections.Generic;
-using System.IO;
-using System.Linq;
-using System.Text.Json;
-using System.Threading.Channels;
-using Telegram.Bot;
-
-namespace AhRulesBot.Infrastructure
-{
-    public static class ServiceCollectionExtensions
-    {
-        internal static IServiceCollection AddTelegramBotClient(this IServiceCollection services)
-        {
-            return services.AddSingleton<ITelegramBotClient>(
-                ctx => new TelegramBotClient(ctx.GetService<AppConfig>().ApiKey));
-        }
-
-        internal static IServiceCollection AddSerilogLogging(this IServiceCollection services)
-        {
-            Log.Logger = new LoggerConfiguration()
-                .MinimumLevel.Debug()
-                .MinimumLevel.Override("Microsoft", LogEventLevel.Information)
-                .MinimumLevel.Override("System", LogEventLevel.Information)
-                .MinimumLevel.Override("Microsoft.AspNetCore", LogEventLevel.Warning)
-                .Enrich.FromLogContext()
-                .WriteTo.Console(new RenderedCompactJsonFormatter())
-                .CreateLogger();
-
-            services.AddSingleton<ILogger>(Log.Logger);
-            return services;
-        }
-
-        internal static IServiceCollection AddAhRulesFile(this IServiceCollection services)
-        {
-            var rules = new List<RuleItem>();
-            var parsed = JsonSerializer.Deserialize<List<Rule>>(File.ReadAllText("rules.json"), new JsonSerializerOptions { PropertyNameCaseInsensitive = true }) ?? new();
-            rules.AddRange(parsed);
-
-            var subrules = parsed.Where(x => x.Rules != null).SelectMany(x => x.Rules).ToList();
-            rules.AddRange(subrules);
-
-            var subsubrules = subrules.Where(x => x.Rules != null).SelectMany(x => x.Rules).ToList();
-            rules.AddRange(subsubrules);
-
-            return services.AddSingleton<List<RuleItem>>(
-                ctx => rules);
-        }
-
-        internal static IServiceCollection AddMsgToRemoveChannel(this IServiceCollection services)
-        {
-            services.AddSingleton<Channel<TelegramMessageInfo>>(Channel.CreateUnbounded<TelegramMessageInfo>(new UnboundedChannelOptions() {
-                SingleReader = true,
-                SingleWriter = true,
-            }));
-            services.AddSingleton<ChannelReader<TelegramMessageInfo>>(svc => svc.GetRequiredService<Channel<TelegramMessageInfo>>().Reader);
-            services.AddSingleton<ChannelWriter<TelegramMessageInfo>>(svc => svc.GetRequiredService<Channel<TelegramMessageInfo>>().Writer);
-            return services;
-        }
-
-        internal static IServiceCollection AddCustomRulesChannel(this IServiceCollection services)
-        {
-            services.AddSingleton<Channel<List<CustomRuleItem>>>(Channel.CreateUnbounded<List<CustomRuleItem>>(new UnboundedChannelOptions()
-            {
-                SingleReader = true,
-                SingleWriter = true,
-            }));
-            services.AddSingleton<ChannelReader<List<CustomRuleItem>>>(svc => svc.GetRequiredService<Channel<List<CustomRuleItem>>>().Reader);
-            services.AddSingleton<ChannelWriter<List<CustomRuleItem>>>(svc => svc.GetRequiredService<Channel<List<CustomRuleItem>>>().Writer);
-            return services;
-        }
-
-        internal static IServiceCollection AddMessageHandlers(this IServiceCollection services)
-        {
-            services.AddSingleton<CustomRulesMessageHandler>(
-                x => new CustomRulesMessageHandler(x.GetRequiredService<ILogger>(), x.GetRequiredService<ChannelReader<List<CustomRuleItem>>>(), null));
-            services.AddSingleton<RulesMessageHandler>(
-                x => new RulesMessageHandler(x.GetRequiredService<ILogger>(), x.GetRequiredService<List<RuleItem>>(), x.GetRequiredService<CustomRulesMessageHandler>()));
-            services.AddSingleton<CommandMessageHandler>(
-                x => new CommandMessageHandler(x.GetRequiredService<RulesMessageHandler>()));
-            services.AddSingleton<UnknownMessageHandler>(
-                x => new UnknownMessageHandler(x.GetRequiredService<CommandMessageHandler>()));
-            services.AddSingleton<IMessageHandler, LongMessageHandler>(
-                x => new LongMessageHandler(x.GetRequiredService<UnknownMessageHandler>()));
-            return services;
-        }
-    }
-}
