+++ src/AhRulesBot/AhRulesBot/Infrastructure/ServiceCollectionExtensions/ServiceCollectionExtensions.cs
@@ -1,0 +1,32 @@
+﻿using Microsoft.Extensions.DependencyInjection;
+using Serilog;
+using Serilog.Events;
+using Serilog.Formatting.Compact;
+using Telegram.Bot;
+
+namespace AhRulesBot.Infrastructure.ServiceCollectionExtensions
+{
+    public static class ServiceCollectionExtensions
+    {
+        internal static IServiceCollection AddTelegramBotClient(this IServiceCollection services)
+        {
+            return services.AddSingleton<ITelegramBotClient>(
+                ctx => new TelegramBotClient(ctx.GetService<AppConfig>().ApiKey));
+        }
+
+        internal static IServiceCollection AddSerilogLogging(this IServiceCollection services)
+        {
+            Log.Logger = new LoggerConfiguration()
+                .MinimumLevel.Debug()
+                .MinimumLevel.Override("Microsoft", LogEventLevel.Information)
+                .MinimumLevel.Override("System", LogEventLevel.Information)
+                .MinimumLevel.Override("Microsoft.AspNetCore", LogEventLevel.Warning)
+                .Enrich.FromLogContext()
+                .WriteTo.Console(new RenderedCompactJsonFormatter())
+                .CreateLogger();
+
+            services.AddSingleton(Log.Logger);
+            return services;
+        }
+    }
+}
